Make *CommandBufferProxy* implement GpuControl
GpuControl is where we will pipe out-of-band stuff, that is currently done in
each of the WGC3D implementations, but we want to move down into
GLES2Implementation.
This is essentially just a refactoring, since the current GpuControl only deals
with GpuMemoryBuffer that's not available out-of-process, but we can then add
things like GenerateMailboxes, Ensure/DiscardBackbuffer or callback stuff on
top.

BUG=181120
R=dmichael@chromium.org, sievers@chromium.org

Review URL: https://codereview.chromium.org/24711002