cc: Update invalidation perftest to clean up released tiles.
The tiles are actually being correctly cleaned up at the end of
the test, but it seems that it accumulates quite a few of them
and times out at the end. This patch ensures that all of the tiles
are cleaned up after each invalidate run.

R=reveman@chromium.org
BUG=298999

Review URL: https://codereview.chromium.org/24487007