BitmapContentLayerUpdater shouldn't reallocate bitmap smaller.
BitmapContentLayerUpdater was always allocating a new bitmap if the size of the region to be drawn changed, even if the new region was smaller. These unnecessary reallocations can be very expensive because they require page faults and for the operating system to clear out the memory.

BUG=296423
R=ben@chromium.org, enne@chromium.org, vangelis@chromium.org

Review URL: https://codereview.chromium.org/24486002