Converted the remaining GPU bots to recipes.
This includes the GPU bots on the chromium.gpu.fyi,
tryserver.chromium, and chromium.webkit waterfalls.

Tested with runit.py/runbuild.py --show-commands against each of the
affected bots, and ensured the correct annotator step was printed.
Also tested the try job path through the GPU recipe locally.

Note that some of the affected bots are in bad shape right now, but
the plan is to work on these bots after the switch.

BUG=286398

Review URL: https://chromiumcodereview.appspot.com/24918002