cc: Add RenderPass::CopyAll().
This method does a deep copy of a RenderPass. This is needed since
we must copy the RenderPass contents from DelegatedRendererLayer to
the DelegatedRendererLayerImpl instead of passing ownership, so that
if the DelegatedRendererLayer moves to a new tree, it can still pass
a copy of the render passes to the impl side again.

Tests:
RenderPassTest.CopyAllShouldBeIdentical

R=piman
BUG=263069
NOTRY=true

Review URL: https://codereview.chromium.org/24619002