Fix flashes when changing tabs.
This is a regression coming in at r225363. Prior to that revision,
the sequence of events when changing a tab was:

  1. Show() the new tab's RWHVMac
  2. WasShown() on the new tab's RWHVMac
     - this called disableScreenUpdatesUntilFlush
  3. Hide() on the old tab's RWHVMac
     - this would call setHidden:YES
  4. WasHidden() on the old tab's RWHVMac
     - this called disableScreenUpdatesUntilFlush

In the new ordering, the sequence is:

  1. Hide() on the old tab's RWHVMac
     - this calls setHidden:YES
  2. WasHidden() on the old tab's RWHVMac
     - this calls disableScreenUpdatesUntilFlush
  3. Show() the new tab's RWHVMac
  4. WasShown() on the new tab's RWHVMac
     - this calls disableScreenUpdatesUntilFlush

Note that in the new ordering, there are no calls to
disableScreenUpdatesUntilFlush until after setHidden is called,
so we get a brief flash of the desktop while the old view is
hidden but the new view is not yet visible.

Revert the part of r225363 that changed window ordering.

Add a call to disableScreenUpdatesUntilFlush in RWHVMac's Hide
function, since the current location in WasHidden is too late to
have the desired effect.

BUG=299858

Review URL: https://codereview.chromium.org/25000005