Fix Device Motion/Orientation browsertest flakyness.
The browser test for Device Motion/Orientation sometimes
fails. This is likely due to a concurrency issue, because
DeviceMotionHostMsg_StopPolling is sent asynchronously to
the browser and there could be a lag on some systems.

This patch fixes the issue by using a WaitableEvent.

BUG=301572
TBR=mvanouwerkerk@chromium.org
NOTRY=true

Review URL: https://codereview.chromium.org/25271002