Pepper: Allow RequestOSFileDescriptor for PRIVATE.
The FileIO_Private API contains the RequestOSFileDescriptor call. However, the
host currently only allows it to be used by whitelisted extensions or if a flag
is passed.

This change also permits the call for plugins with PERMISSION_PRIVATE. This
allows this interface to be used by the NaCl "trusted plugin".

BUG=246396
R=bbudge@chromium.org

Review URL: https://codereview.chromium.org/25049005