Files.app: Remove ProgressCenterClient class.
Originally ProgressCenter in the background page and application window
communicates by using chrome.message. But the other components in Files.app such
as FileOperationManager does not uses the message for the communication and
calls functions directly. For consistency, this CL stops to use the message for
the communication and remove the client helper of ProgressCenter, which is no
longer needed after stopping to use chrome.message.

BUG=277432
TEST=manually

Review URL: https://codereview.chromium.org/25226003