[chromium_android recipes] Add intrumentation tests recipes.
The internal_dir representation in the chomium_android module
is now a config param of type Path which depends on the
dynamic path '[CHECKOUT]'. Hence, this CL removes the special
case to handle the earlier representation of internal_dir
(which was evaluated as a static value based on path.checkout) in the path module.

BUG=286509

Review URL: https://chromiumcodereview.appspot.com/24530005