ChildProcessSecurityPolicy: Remove CanWriteFile and CanCreateFile.
We currently have some checks that are more fine-grained than the grants. There's no reason for this, so I'm removing them and replacing them with the broader check.

I would also like to remove some of the Can*FileSystemFile checks, but I currently cannot, as that goes into checking the permissions for the filesystem as a whole in the case of isolated file systems.

That will have to be saved for a different CL.

This CL depends on:
https://codereview.chromium.org/24406003/

This is part of Step 2 of this refactoring plan:
https://docs.google.com/a/google.com/document/d/1QGkGWuwgSuaRqovz4wyb0upqPKDVsgYOFKt44E7gmOE/edit?usp=sharing

BUG=262142, 263150

Review URL: https://codereview.chromium.org/24439002