Fix use-after-free on base::StatsCounter::GetPtr.
ChromeMainDelegate has 2 objects used for timing startup: startup_timer_ and
stats_counter_timer_.

However, the former is used as a wrapper around a
reference of the latter. The destructors of these 2 objects are getting called
in the wrong order such that we have a use-after-free error when startup_timer_
tries to call Stop() on stats_counter_timer_.

This patch switches the declaration order of the two member variables to ensure
that they are cleaned up in the right order.

BUG=292937
R=cpu@chromium.org, jar@chromium.org

Review URL: https://codereview.chromium.org/24255007