Copy 'patch_url' property from the parent to the triggered builder.
When a builder is triggered by a 'parent' builder, which was
in turn triggered by an SVN try job, then the triggered
'child' builder will also need to apply the patch which
triggered the 'parent'. For example, if the 'child' were a
tester which downloads the build product of the 'parent',
then the 'child' tester would need the patch as the patch
could potentially have data files required for the tests.

Review URL: https://chromiumcodereview.appspot.com/25532002