[Android] Only vibrate when selection handles are actually shown
Previously, we were vibrating when there was a new selection and
selection handles weren't already shown. There are, however, some cases
where this selection won't trigger the handles to be shown. In that
state, a zoom/scroll would trigger vibrate on every frame.

Instead, we should only be vibrating when the handles go from not-shown
to shown.

BUG=301533,150151

Review URL: https://codereview.chromium.org/25301002