Fix crash when ECHO_TO_CONSOLE
TraceEvent's copy constructor has a side-effect that the ownership
of the convertable values will be passed from the source TraceEvent
to the target TraceEvent.

Handle ECHO_TO_CONSOLE before the event is added to the buffer.

The copy constructor will be obsoleted soon to avoid such issues.

BUG=302676
TEST=Run cc_unittests with --trace_to_console

Review URL: https://codereview.chromium.org/25453005