Always tear down the OSX app list signin view, when the signin status changes.
This improves a fix in r218251, after which the signinDelegate on the
signin view would not be replaced when the model changes, but when
both models were signed out.

BUG=297437
TEST=Have the OSX App Launcher enabled and a single, signed out profile.
Delete the profile from chrome://settings, then click sign in on the app
launcher. Chrome should not crash.

Review URL: https://codereview.chromium.org/25649004