aura: Use Layer::SetShowPaintedContent to stop showing external content
Currently we use SetExternalTexture(NULL) or SetDelegatedFrame(NULL)
to switch back to showing painted content. This is kinda indirect,
and won't play nicely when SetExternalTexture() is removed, and when
we use a FrameProvider for delegated frames instead of pushing them
to the ui::Layer directly one at a time.

Instead, add SetShowPaintedContent() to switch to a content layer,
and disallow setting a NULL/empty texture or delegated frame.

Also View::SetExternalTexture() is not used, remove it so it doesn't
have to think about this change.

R=piman,sadrul
BUG=263069

Review URL: https://codereview.chromium.org/25596002