net: remove TLS fallback for Google properties.
With Chrome 31, Chrome will not accept a fallback to SSLv3 for Google
properties. If we wish to be able to depend on AES-GCM and ChaCha20+Poly1305
support then we need to ensure that an attacker cannot downgrade us to TLS 1.1
(which doesn't include those options).

This change prototypes a signal to indicate that fallback should
not be allowed. We know that Google properties don't need fallback and this
change allows us to test that networks are not too broken to deploy such a
scheme.

It's possible that defective MITM proxies will cause issues with this. The
--enable-unrestricted-ssl3-fallback option exists for diagnosis of any bug
reports.

BUG=302739

Review URL: https://codereview.chromium.org/25547003