[NaCl SDK] nacl_io: Re-enable TCP test
The adds and in-process PPAPI echo server to
nacl_io_socket_test.  Doing it this way rather then
relying on an external server, or a server in thttpd.py
means the tests are more self-contained and can run in
packages form as well.

Also, refactored some of the test fixtures to use SetUp
and TearDown rather than constructors and destructors.

R=binji@chromium.org, noelallen@chromium.org

Review URL: https://codereview.chromium.org/25350004