[NaCl Docs] Add a new doc build configuration: devsite-staging.
In this mode the html files expect to live under a folder
called $USER which allows each of us to stage our own copy
of the docs.

In order to implement this I added a new devsite config
option called devsite_foldername which defaults to 'dev'.

R=eliben@chromium.org

Review URL: https://codereview.chromium.org/25548007