Separate PepperPluginRegistry initialization from constructor.
When initializing the registry, we initialize all internal modules. If a module
fails to initialize, it is destroyed before the registry initialization
finishes, but the PluginModule destructor calls back into
PepperPluginRegistry::GetInstance().
As it is now, this tries to recreate the registry, causing an infinite
recursion. So instead, create the registry first, and then initialize it so that
module initialization can safely call PepperPluginRegistry::GetInstance().

BUG=None

Review URL: https://codereview.chromium.org/25569004