Makes RecentlyUsedFoldersComboModel clear internal structur when node deleted
Prior to this RecentlyUsedFoldersComboModel would not clear its
internal structures if a node was deleted, leading to a future
crash. This patch makes RecentlyUsedFoldersComboModel clear internal
structure and notify observers.

Having the BookmarkModel change while the bubble is showing is rare,
so the code doesn't attempt to do anything elegant.

BUG=294330
TEST=none
R=msw@chromium.org

Review URL: https://codereview.chromium.org/25310003