Add "Bookmarks File" to ImporterList interactively.
Otherwise it becomes an option on first run AutoImport and that is not desired. 

Instead we force the clients to make the decision on whether they want or not to add the "interactive profile", aka "Bookmarks HTML File" to the ImporterList, which can make it available in the "Import Bookmarks and Settings" dialog.

BUG=295755
TEST=launch chrome, go to chrome://settings/importData, verifies that
"Bookmarks HTML file" option is present in the "From:" combobox and still works.
R=gab@chromium.org

Review URL: https://codereview.chromium.org/24345002