Don't override optional audio constraints with defaults.
The app should be able to set media constraints in either the mandatory
or optional fields and have them respected.

TESTED=Using apprtc with this change to allow passing constraints:
http://review.webrtc.org/2345004/
verified that constraints in both mandatory and optional fields are
not overridden by defaults.

Review URL: https://codereview.chromium.org/25683007