sync: Fix bug in early sync shutdown
In r224014, we tried to fix many issues with early shutdown.  That CL
contained a fairly serious bug that would cause shutdown requests that
arrive very early to trigger a crash.

This patch fixes that bug and adds a test to help prevent regressions.

BUG=236451,303177

Review URL: https://codereview.chromium.org/25794002