[rAC, OSX] Make sure sign in view renders.
Since we enable auto-resize, the WebCore FrameView demands a non-empty
minimum size, which is requested from the delegate. Since the delegate
was initialized in loadView, it had an empty min and max size, causing
the sign-in contents to be invisible.

This change constructs the delegate when showing the sign-in dialog
instead, ensuring minimum and maximum size are valid.

BUG=none
R=isherman@chromium.org

Review URL: https://codereview.chromium.org/25658002