Avoid crash when playing corrupt audio
Corrupt audio data in a media file can cause a crash if the number of channels
exceeds the limit that Chrome supports. Fix is to return an error if the number
of channels is too large, so that the buffer allocation is not attempted.

BUG=301985
TEST=Video now plays without crashing (there is no sound from the corrupt audio portion)

Review URL: https://codereview.chromium.org/25772003