Rework the test shutdown order a little.
It's illegal (and soon to be a runtime error) to touch Blink objects after
calling WebKit::shutdown(). This CL reworks destruction order a little to
make sure that does not happen.

It also has the pleasing symmetry that WebKit::initialize and ::shutdown
are now nicely paired in the constructor and destructor of the test
support class.

BUG=302754

Review URL: https://codereview.chromium.org/25629005