Disable UI controls for maximize and fullscreen if an app window is constrained. (Mac)
This prevents an app window from being maximized or fullscreened by the
user if it has a maximum width or height, or if it is not resizable. Maximize
or fullscreen via the window API is still allowed.

Also, a window is not resizable if it's min and max sizes are specified
and equal.

BUG=168080,261130,242678

Review URL: https://codereview.chromium.org/23856020