Ensure that in Windows AURA in non DWM mode modal dialogs get keyboard/mouse input when the main window is activated.
In non DWM(Desktop Window manager) mode the modal dialogs are created as Widgets backed by NativeWidgetAura instances which
means that they are not top level windows. When the window which owns these dialogs loses activation, and we click on this
window then WM_MOUSEACTIVATE for the top level window needs to return MA_ACTIVATE which means that the window can be activated.

Currently we return MA_NOACTIVATEANDEAT in this codepath when a modal dialog is active. For top level HWND based modal dialogs
the owner window is disabled and hence it never receives a mouse activate when we click on the window which is why this bug
was not seen in regular Windows Chrome or DWM Windows AURA. To get this to work correctly we need to return MA_ACTIVATE when
a modal window is active.

Added a function IsModalWindowActive to the DesktopRootWindowHostWin class. This is invoked during activation checks.
We also need to override the BaseFocusRules::GetTopLevelWindow function in the DesktopFocusRules class to return the
content_window_ in case we are trying to find the window to activate and arrive at the dummy container window during
enumeration.

BUG=303207
TEST=Covered by windows AURA views unittest WindowModalityActivationTest
R=ben@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/25941002