[chromium_android] Comment out logcat monitor and its dependents.
We would like to do this now so that we can test the rest of the
API while we discuss ways to spawn background processes from
recipes. I have added a TODO with each commented block to
uncomment them once the logcat monitor is enabled.

BUG=286509

Review URL: https://chromiumcodereview.appspot.com/26030003