cc: Fix hit-testing of empty layers.
The RenderSurfaceLayerList is supposed to contain any layers which are visible
or that are eligible for hit testing. This change ensures that empty layers
(those with drawsContent=false) are left in the RSSL when they have touch or
wheel handlers, so that we can find them in hit testing.

BUG=289766
R=danakj@chromium.org

Review URL: https://codereview.chromium.org/25712004