Check if RTI::SharedMainThreadContextProvider is there before wrapping
It's not valid to wrap a NULL cc::ContextProvider in a
WebGraphicsContext3DProvider, so NULL check it before wrapping. The
SharedMainThreadContextProvider() could be NULL if we can't create
and bind a valid context, which can happen if the GPU is unhappy or
the process crashes any early point in initialization.

BUG=301555

Review URL: https://codereview.chromium.org/26028004