Add a "Learn more" link in the "Sign in with a different account" confirmation dialog.
Add a "Learn more" link to a troubleshooting page at the bottom of the "Sign in with a different account" dialog.

Expected result:
http://i.imgur.com/txdpn4R.png (Mac)
http://i.imgur.com/QvIfDr4.png (Windows)

TEST=
1. Sign in to Chrome with a first account (hot-dog menu > Sign in to Chrome...).
2. Sign out (hot-dog menu > Settings > Disconnect your Google Account...).
3. Sign in to Chrome with a different account than the one used at step 1.
Expected result: In the "Sign in with a different account" confirmation dialog that appears, there should be a "Learn more" link.
4. Click on the "Learn more" link.
Expected result: A help page is loaded in a popup.
BUG=268379

Review URL: https://codereview.chromium.org/22165004