Fix misuse of STL in InvalidatorRegistrar
The invocation of std::set_intersection() in InvalidatorRegistrar was
using the wrong kind of output iterator, which would lead to crashes
under some conditions (eg. valgrind).  It was technically invalid under
any conditions; we were "lucky" that it didn't crash most of the time.

BUG=304250

Review URL: https://codereview.chromium.org/25855007