Remove dead IPC WorkerMsg_PostMessage.
Some places claimed to send this message, but nobody claimed to receive it, so
it's dead.

When the IPC was originally added in https://codereview.chromium.org/160576,
there was a receiver. The something changed, and the file containing the
receiver was removed by http://codereview.chromium.org/8416029.

This might be related to the fact that we cannot do myWorker = new
SharedWorker(); myWorker.postMessage() (but we need to do
myWorker.port.postMessage()).

BUG=NONE
R=atwilson@chromium.org, dslomov@chromium.org, jochen@chromium.org, tsepez@chromium.org
TBR=tsepez

Review URL: https://codereview.chromium.org/25674003