[Activity Log] Enable the webrequest tests on windows.
Previously they were logging different activity. The recent fixes we have done to the tests have also fixed this.  They are now logging the same as on linux and mac so this change deletes the win specific override and enables the test on win.


BUG=292252

Review URL: https://codereview.chromium.org/26090003