Require results directory when generating JSON output in runtest.py.
An earlier CL attempted to make the results directory optional, but
too much code in the flakiness dashboard upload code requires it to be
present. Instead of crashing with a Python exception, make it an error
to specify --generate-json-file=True to runtest.py without also
specifying --results-directory.

If the target test doesn't generate an XML file then its log will be
parsed and uploaded.

Updated the GPU recipe to set this argument for the Telemetry based
WebGL conformance tests.

This has been tested as much as possible locally without actually
uploading bogus results to the flakiness dashboard.

BUG=225939

Review URL: https://chromiumcodereview.appspot.com/26331002