Fix regression where all checkboxes are disabled on "Import Bookmarks ..." dialog.
This regression happened because of a typo committed in r224933, at line 235
of this file. In that revision |state| parameter was renamed to |importing| in
setImportingState() function for clarity, but not all references were updated,
we have missed one when setting the disable state of the checkboxes.

BUG=299508
TEST=
1.Launch chrome and navigate to chrome://settings.
2.Click on 'Import bookmark and settings' button under 'Users' section.
3.Select 'Mozilla firefox' from dropdown and click On 'Import button'.
4.Now Click on 'Skip import' button.
5.Repeat step 2 and Observe the checkboxes.
They were "disabled", now this fixes the behavior to make them "enabled" again.

R=dubroy@chromium.org

Review URL: https://codereview.chromium.org/26123002