Fix service process not being launched in split dll
chrome.exe was using the existence of "--type" to determine which DLL
to load, but ServiceProcessMain is in the browser DLL, not the child,
so when launched with --type=service the result is an immediate exit.

This broke cloud print (and probably? bookmark import).

Explicitly check for this process type to force the browser DLL to
be loaded.

R=vitalybuka@chromium.org
TBR=cpu@chromium.org
BUG=304715,237249

Review URL: https://codereview.chromium.org/26306004