Add ComboboxModel::IsItemEnabledAt
This CL enables to switch the enabling/disabling state of menu item of a Combobox.

Concretely, this feature is needed by new Translate UI. Actually, the Translate current infobar has this feature. For example, the Translate UI has the source language and the target language selectors as Comboboxes, and when English is selected as the source language, English in the target language selector should be disabled.

BUG=276181

Review URL: https://codereview.chromium.org/26219002