Android WebView: post QuotaManager calls to UI thread.
The AwQuotaManagerBridge needs to be usable from any thread, not just
the UI thread. Make it refcounted and have its methods post to the UI
thread if not already there. Since none of the methods are synchronous
there's no need to wait for the posted messages to be handled.

BUG=305205
R=benm@chromium.org

Review URL: https://codereview.chromium.org/26477002