Add Clang plugin check for WeakPtrFactory member order
Also consolidates some of the FindBadConstruct options in to a small struct for readability.

The WeakPtrFactory member order is behind a flag, because
there are many violations in Chromium right now. Also, there
might need to be exceptions in the checker. For example,
we might want to allow more than one WeakPtrFactory to refer
to the outer class, as long as all such factories appear
after all other members. That's TBD.

BUG=303818
R=erg@chromium.org

Review URL: https://codereview.chromium.org/26303002