Destroy references to context3d_ before leaking it and setting to NULL.
The ContextProviderCommandBuffer callback proxies make use of the
provider's context3d_ when they are destroyed. If the context is to
be leaked on shutdown however, we leak the context3d_ and set it
to NULL in the destructor. Then the callback proxies will try to use
a NULL pointer. So, destroy them before leaking the context3d_.

Tests:
ContextProviderCommandBufferBrowserTest.LeakOnDestroy

R=piman
BUG=304861

Review URL: https://codereview.chromium.org/26182004