Fix win_nacl_sdk_build builder.
On windows we need to add python.exe as the first argument.
All the annotator scripts are currently python so there
shouldn't be any negative effects to doing this in all
cases.

AddAnnotatedSteps already does this so this simply brings
AddAnnotationStep into parity with it.

Review URL: https://codereview.chromium.org/25888003