Disable image transport surface on all windows machines.
We want to try using the direct presentation path everywhere, so blacklist texture sharing on all machines. If this works, we can delete the image transport surface code later.
We've made the image transport surface work even with the sandbox at full strength, so remove the code that reduces the GPU sandbox in this case.

BUG=304369
R=kbr@chromium.org

Review URL: https://codereview.chromium.org/26073002