Ensure that in desktop AURA if the dummy container window does not inadvarently get activated because of a sibling window (modal window) getting destroyed.
We should activate its child window which hosts the real content in this case.

Added an implementation for the DesktopRootWindowHostX11::InitFocus function to set the focused window in the focus
client. The DesktopRootWindowHostX11 class was setting the focused window in the InitRootWindow function which is early
because the content window being focused is not yet visible.

Fixes bug https://code.google.com/p/chromium/issues/detail?id=303484

BUG=303484
R=sky@chromium.org, sky
TEST=Covered by interactive ui test WindowModalWindowDestroyedActivationTest

Review URL: https://codereview.chromium.org/26081002