Add the concept of a source set to GN.
A source set acts like a static library in most cases, but doesn't actually make the library. The object files will be added to targets that depend on it.

I changed the default component mode to be a "source set" rather than a shared library in non-component mode.

This fixes the script target writer unit test on Windows.

BUG=
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/26267003