Linux sandbox: Allow the GPU process to go past 4GB
The GPU process is currently limited to 4GB of address space,
which causes crashes in some legitimate cases (the canonical
example is to open a ton of empty tabs on Chrome OS, when the
GPU process's virtual address space reaches 4GB it crashes).

This didn't use to be a problem for Chrome OS before we had
compressed swap as we were limited to 4GB of physical memory,
but now that we have compressed swap and that a lot of Chrome OS
machines are switching to 64 bit, it is possible to get there
without problems.

To fix this, we bump the maximum size of the GPU process's
address space to 16GB on 64 bit.

BUG=chromium:271119
TEST=open a ton of tabs, see VIRT go beyond 4096M in top and the GPU process doesn't crash

Review URL: https://codereview.chromium.org/26102002