Reland r227565 "Enable EncryptedMediaTest (in content_browsertest) on Android"
This reverts r227627 with fix.

Encrytped media playback with <video src=...> is not supported on Android.
Those tests will not run on Android.

There tests should not run on devices where MSE/EME are not available. We need
to find a way to fix that before we can enable them by default.

BUG=304956, 300035
TBR=ddorwin@chromium.org

Review URL: https://codereview.chromium.org/26574004