In ExtensionBrowserTest::ReloadExtension, wait for the background page to execute before returning.
We do this in LoadExtension and InstallOrUpdateExtension, and
missing it in ReloadExtension can cause hard-to-understand or
possibly even flaky failures.  The extra call does nothing in the
case that there aren't any loading views.

I didn't add this to EnableExtension because that doesn't run the
message loop at all, so the failure to wait for views to load is
less surprising.

Review URL: https://codereview.chromium.org/26451005