[InfoBar] Always close native infobars from C++
We have an ad-hoc mechanism to close infobars from java on navigations. That is only needed for java only infobars which will hopefully dissappear soon.

Also remove the CloseinfoBar method which closed the java counterpart first. It's better to be consistent with how C++ does it and always close the C++ piece first.

BUG=283720
TEST=All downstream tests keep passing
     Manual testing of the scenarios affected

Review URL: https://codereview.chromium.org/26176007