Fix bug in calculating minimize button position
When DWM is enabled it appears that WM_GETTITLEBARINFOEX returns
values in terms of the classic theme and only returns the right thing
after an activate. This is problematic for tab dragging code as it
needs to know the size of the tabstrip before the window has shown.

The fix is to cache the offset and use the cached value until we get
the activate.

BUG=293698
TEST=none
R=ananta@chromium.org

Review URL: https://codereview.chromium.org/26518005