Create an app_shell target
The goal of app_shell is to depend only on content (not chrome), and run
an app instead of a web browser. This is the initial build target. For now,
only Chrome OS is supported. The app initializes an X window, an Aura
RootWindow, and creates a sample views::Widget containing a WebView to
demonstrate that src/content is properly initialized.

BUG=297890
TEST=Run app_shell on linux_chromeos, see google.com load in a window
R=derat@chromium.org, miket@chromium.org

Review URL: https://codereview.chromium.org/26603002