Explicit initialization of aura::Env for browser shell.
Most implementations (x11 aura with and without chromeos) do not explicitly
call aura::Env::GetInstance() to initialize it. These build work as is because
the call is done when performing other operations, such as adding observers.

In OZONE there are no such calls, so the OZONE build fails when trying to
create a compositor since the Compositor::Initialize hasn't been called.

This patch adds an explicit initialization for aura based builds to fix this
issue.

BUG=

Review URL: https://codereview.chromium.org/23882007